+++ Sources.cmake
@@ -1,8 +1,8 @@
-target_sources(Queijo.Runtime PRIVATE
-    runtime/include/queijo/options.h
-    runtime/include/queijo/ref.h
-    runtime/include/queijo/require.h
-    runtime/include/queijo/runtime.h
+target_sources(Lute.Runtime PRIVATE
+    runtime/include/lute/options.h
+    runtime/include/lute/ref.h
+    runtime/include/lute/require.h
+    runtime/include/lute/runtime.h
 
     runtime/src/options.cpp
     runtime/src/ref.cpp
@@ -10,39 +10,39 @@ target_sources(Queijo.Runtime PRIVATE
     runtime/src/runtime.cpp
 )
 
-target_sources(Queijo.Fs PRIVATE
-    fs/include/queijo/fs.h
+target_sources(Lute.Fs PRIVATE
+    fs/include/lute/fs.h
 
     fs/src/fs.cpp
 )
 
-target_sources(Queijo.Luau PRIVATE
-    luau/include/queijo/luau.h
+target_sources(Lute.Luau PRIVATE
+    luau/include/lute/luau.h
 
     luau/src/luau.cpp
 )
 
-target_sources(Queijo.Net PRIVATE
-    net/include/queijo/net.h
+target_sources(Lute.Net PRIVATE
+    net/include/lute/net.h
 
     net/src/net.cpp
 )
 
-target_sources(Queijo.Task PRIVATE
-    task/include/queijo/task.h
+target_sources(Lute.Task PRIVATE
+    task/include/lute/task.h
 
     task/src/task.cpp
 )
 
-target_sources(Queijo.VM PRIVATE
-    vm/include/queijo/spawn.h
-    vm/include/queijo/vm.h
+target_sources(Lute.VM PRIVATE
+    vm/include/lute/spawn.h
+    vm/include/lute/vm.h
 
     vm/src/spawn.cpp
     vm/src/vm.cpp
 )
 
-target_sources(Queijo.CLI PRIVATE
+target_sources(Lute.CLI PRIVATE
     cli/main.cpp
     cli/tc.h
     cli/tc.cpp
