+++ extern/luau/Ast/include/Luau/Parser.h
@@ -62,7 +62,7 @@ public:
         ParseOptions options = ParseOptions()
     );
 
-// FIXME(queijo): temporary change, we should expose a few more methods here ultimately
+// FIXME(lute): temporary change, we should expose a few more methods here ultimately
 //private:
     struct Name;
     struct Binding;
