+++ runtime/src/options.cpp
@@ -1,4 +1,4 @@
-#include "queijo/options.h"
+#include "lute/options.h"
 
 // TODO: this is never set to true today
 static bool codegen = false;
